Computer_Vision_Project/requirements.txt
numpy
opencv-python
ultralytics
easyocr
pyyaml
pytest

